requirements.txt
# Core dependencies
wheel
python-dotenv
requests
pandas-ta
numpy
pandas
finta

# Web framework
flask
flask-socketio
python-engineio
python-socketio
eventlet
websockets
websocket-client
flask-sqlalchemy
flask-login
gunicorn

# Trading and exchanges
ccxt
schedule

# Solana and blockchain
solana
solders
base58
borsh-construct
pynacl
cryptography
anchorpy

# Async and networking
aiohttp
asyncio
httpx
aiosqlite

# Data analysis and visualization
matplotlib
scipy
scikit-learn
seaborn
plotly
dash

# Database
sqlalchemy

# Utilities
retrying
pytest
twikit
statistics
backoff
FastAPI 

# Settings management
pydantic
pydantic-settings
cachetools